refactor(api): restructure events with domain files and structured types
Split events.rs into events/block.rs, events/player.rs, events/chat.rs.
Events use structured types: BlockPosition, Position, Rotation,
ChunkPosition. Player identity removed from all events — use
ctx.player() instead. PlayerJoinedEvent and PlayerLeftEvent are now
unit structs. Updated all plugins and basalt-server dispatch sites.
